--- core/nodes/index.ts
@@ -1,2 +1,0 @@
-export * from "./types";
-export * as light from "./lignt";
