--- core/nodes/lignt/index.ts
@@ -1,2 +1,0 @@
-export { makeReference } from "../types/reflect-node-reference";
-export type { IReflectNodeReference } from "../types/reflect-node-reference";
