+++ core/utils/check-if-auto-layout.ts
@@ -1,7 +1,7 @@
 import { BaseFrameMixin } from "@design-sdk/figma";
 import { convertLayoutModeToAxis } from "@design-sdk/figma-node-conversion";
 import { Axis } from "@reflect-ui/core";
-import { ReflectFrameNode } from "../nodes";
+import { ReflectFrameNode } from "@design-sdk/figma-node";
 
 export function checkIfAutoLayout(
   node: ReflectFrameNode | BaseFrameMixin
