+++ core/utils/common-padding.ts
@@ -1,4 +1,4 @@
-import { ReflectSceneNode } from "@design-sdk/core/nodes";
+import { ReflectSceneNode } from "@design-sdk/core";
 import { EdgeInsets } from "@reflect-ui/core";
 
 type Padding = {
