+++ figma-node-conversion/conversion/force-convert-rect-to-frame.ts
@@ -1,4 +1,4 @@
-import { ReflectFrameNode, ReflectRectangleNode } from "@design-sdk/core/nodes";
+import { ReflectFrameNode, ReflectRectangleNode } from "@design-sdk/core";
 import { CrossAxisAlignment, MainAxisAlignment } from "@reflect-ui/core";
 
 // TODO - Additional feature, this needs to be migrated to @designto/token/logics
