+++ figma-node-conversion/converters/corner-radius.convert.ts
@@ -1,5 +1,5 @@
 import { BorderRadiusManifest } from "@reflect-ui/core";
-import { mixed } from "@design-sdk/core/nodes";
+import { mixed } from "@design-sdk/core";
 import { FigmaCornerRadius } from "@design-sdk/figma-types";
 
 export function convertFigmaCornerRadiusToBorderRadius(
