+++ figma-reflect-node/README.md
@@ -1,0 +1,3 @@
+# Figma Reflect Node Alias (Temporary)
+
+> Currently, Reflect design nodes rely on figma node definitions. This will be migrated to native reflect-ui core definitions in the future. Meanwhile, Reflect node definitions are based on figma and this is the main node definitions package.
