+++ figma-reflect-node/index.ts
@@ -13,24 +13,5 @@ export * from "./text.node";
 export * from "./vector.node";
 export * from "./mixed";
 
-export // mixed,
-// // abstract
-// ReflectSceneNode,
-// ReflectSceneNodeType,
-// ReflectBaseNode,
-// ReflectChildrenMixin,
-// // nodes
-// ReflectTextNode,
-// ReflectConstraintMixin,
-// ReflectEllipseNode,
-// ReflectRectangleNode,
-// ReflectFrameNode,
-// ReflectGroupNode,
-// ReflectLineNode,
-// mixin
-// ReflectBlendMixin,
-// ReflectCornerMixin,
-// ReflectDefaultShapeMixin,
-// ReflectGeometryMixin,
-// ReflectLayoutMixin,
- {};
+// reference nodes
+export * from "./reflect-node-reference";
