+++ figma-types/lib/corner-radius.ts
@@ -1,4 +1,4 @@
-import { mixed } from "@design-sdk/core/nodes";
+import { mixed } from "@design-sdk/core";
 
 export interface FigmaCornerRadius {
   cornerRadius: number | typeof mixed;
