+++ vanilla/make/cgrect.make.ts
@@ -1,4 +1,4 @@
-import { ReflectFrameNode, ReflectRectangleNode } from "@design-sdk/core/nodes";
+import { ReflectFrameNode, ReflectRectangleNode } from "@design-sdk/core";
 import { CGRectManifest } from "@reflect-ui/core/lib";
 
 export function makeCGRect(
